Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -19,7 +19,7 @@ int main()
         case '/':
             if(b!=0)
             {
-            printf("%d\n",a/b);
+            printf("%f\n",a/(float)b);
             }
             else 
             {
